mcp-servers/postgres-query-server/requirements.txt
mcp>=1.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
